update exception, html, entities

+++ MainProject/QuanLyThuVienNhacOnline/web/src/main/java/com/musicmanager/web/service/ArtistService.java
@@ -2,6 +2,8 @@ package com.musicmanager.web.service;
 
 import com.musicmanager.web.entity.Artist;
 import com.musicmanager.web.entity.Song;
+import com.musicmanager.web.exception.BadRequestException;
+import com.musicmanager.web.exception.ResourceNotFoundException;
 import com.musicmanager.web.repository.ArtistRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,10 +24,9 @@ public class ArtistService
     }
 
     // LẤY THÔNG TIN CỦA NGHỆ SĨ THEO ID
-    // TODO: THÊM EXCEPTION
     public Artist getArtist(String id)
     {
-        return artistRepository.findById(id).orElse(null);
+        return artistRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Artist with id: " + id + " not found!."));
     }
 
     // TÌM NGHỆ SĨ THEO TÊN (ĐẠI KHÁI)
@@ -40,50 +41,50 @@ public class ArtistService
         return artistRepository.findByNameIgnoreCase(name);
     }
 
-    // LẤY DANH SÁCH BÀI HÁT CỦA NGHỆ SĨ
-    public List<Song> getArtistSongs(String id)
-    {
-        Artist artist = getArtist(id);
-
-        if (artist != null)
-        {
-            return artist.getSongs();
-        }
-
-        return Collections.emptyList();
-    }
-
     // TẠO NGHỆ SĨ MỚI
-    // TODO: THÊM EXCEPTION
     public void createArtist(Artist request)
     {
-        if (searchArtistByName(request.getName()) == null)
+        if (request.getName() == null || request.getName().isBlank())
         {
-            artistRepository.save(request);
+            throw new BadRequestException("Artist name cannot be blank!.");
         }
+
+        if (searchArtistByName(request.getName()) != null)
+        {
+            throw new BadRequestException("Artist name existed!.");
+        }
+
+        artistRepository.save(request);
+
     }
 
     // SỬA THÔNG TIN NGHỆ SĨ
-    // TODO: THÊM EXCEPTION
     public void updateArtist(String id, Artist request)
     {
         Artist artist = getArtist(id);
-        Artist existingArtist = searchArtistByName(request.getName());
-        if (existingArtist == null || existingArtist.getId().equals(id))
-        {
-            artist.setName(request.getName());
-            artist.setCountry(request.getCountry());
-            artist.setGender(request.getGender());
-            artist.setInformations(request.getInformations());
 
-            artistRepository.save(artist);
+        Artist existingArtist = searchArtistByName(request.getName());
+        if (existingArtist != null && !existingArtist.getId().equals(id))
+        {
+            throw new BadRequestException("Artist name existed!.");
         }
+
+        if (request.getName() != null) artist.setName(request.getName());
+        if (request.getCountry() != null) artist.setCountry(request.getCountry());
+        if (request.getGender() != null) artist.setGender(request.getGender());
+        if (request.getInformations() != null) artist.setInformations(request.getInformations());
+
+        artistRepository.save(artist);
     }
 
     // XÓA NGHỆ SĨ
-    // TODO: THÊM EXCEPTION
     public void deleteArtist(String id)
     {
+        if (!artistRepository.existsById(id))
+        {
+            throw new ResourceNotFoundException("Cannot find Artist with id: " + id);
+        }
+
         artistRepository.deleteById(id);
     }
 
@@ -91,12 +92,5 @@ public class ArtistService
     public long numberOfArtists()
     {
         return artistRepository.count();
-    }
-
-    // SỐ LƯỢNG BÀI HÁT CỦA NGHỆ SĨ
-    // TODO: THÊM EXCEPTION
-    public long numberOfArtistSongs(String id)
-    {
-        return getArtistSongs(id).size();
     }
 }
